Zmiana typu wartości temperatury na double. Wstępna implementacja urządzenia grzewczego.

+++ OfficeHVAC.Modules.RoomSimulator/ViewModels/IRoomViewModel.cs
@@ -2,6 +2,6 @@
 {
     public interface IRoomViewModel
     {
-        float Temperature { get; set; }
+        double Temperature { get; set; }
     }
 }
